+++ modules/juce_audio_utils/players/juce_AudioProcessorPlayer.cpp
@@ -153,7 +153,7 @@ void AudioProcessorPlayer::audioDeviceAboutToStart (AudioIODevice* const device)
     numOutputChans = numChansOut;
 
     messageCollector.reset (sampleRate);
-    channels.calloc (jmax (numChansIn, numChansOut) + 2);
+    channels.calloc ((size_t) jmax (numChansIn, numChansOut) + 2);
 
     if (processor != nullptr)
     {
